+++ pkgs/jnigen/CHANGELOG.md
@@ -12,6 +12,7 @@
 - **Breaking Change**: Removed the `Impl` suffix from the generated
   implemenation classes. So the implementation class for an interface named
   `Foo` is now simply called `$Foo` instead of `$FooImpl`.
+- **Breaking Change**: Renamed `castTo` to `as`.
 - Added `JImplementer` which enables building an object that implements multiple
   Java interfaces. Each interface now has a static `implementIn` method that
   takes a `JImplementer` and the implementation object.